Split cpp code from the Python comment && Add a little refactor to convert
